requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.0
django-allauth==0.44.0
django-currentuser==0.6.1
django-filter==23.5
django-jazzmin==2.6.0
django-sendsms==0.5
django-smart-selects==1.6.0
django-translation==1.2
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
twilio==8.11.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
